LWE: refactor, scale-related verifier and support level reduce
