+++ lib/Dialect/BGV/IR/BGVOps.cpp
@@ -1,0 +1,67 @@
+#include "lib/Dialect/BGV/IR/BGVOps.h"
+
+#include "lib/Dialect/LWE/IR/LWEOps.h"
+
+namespace mlir {
+namespace heir {
+namespace bgv {
+
+//===----------------------------------------------------------------------===//
+// Op verifiers
+//===----------------------------------------------------------------------===//
+
+LogicalResult MulOp::verify() { return lwe::verifyMulOp(this); }
+
+LogicalResult MulPlainOp::verify() { return lwe::verifyMulPlainOp(this); }
+
+LogicalResult RotateColumnsOp::verify() { return lwe::verifyRotateOp(this); }
+
+LogicalResult RotateRowsOp::verify() { return lwe::verifyRotateOp(this); }
+
+LogicalResult RelinearizeOp::verify() { return lwe::verifyRelinearizeOp(this); }
+
+LogicalResult ModulusSwitchOp::verify() {
+  return lwe::verifyModulusSwitchOrRescaleOp(this);
+}
+
+LogicalResult ExtractOp::verify() { return lwe::verifyExtractOp(this); }
+
+LogicalResult LevelReduceOp::verify() { return lwe::verifyLevelReduceOp(this); }
+
+//===----------------------------------------------------------------------===//
+// Op type inference.
+//===----------------------------------------------------------------------===//
+
+LogicalResult AddOp::inferReturnTypes(
+    MLIRContext *ctx, std::optional<Location>, AddOp::Adaptor adaptor,
+    SmallVectorImpl<Type> &inferredReturnTypes) {
+  return lwe::inferAddOpReturnTypes(ctx, adaptor, inferredReturnTypes);
+}
+
+LogicalResult SubOp::inferReturnTypes(
+    MLIRContext *ctx, std::optional<Location>, SubOp::Adaptor adaptor,
+    SmallVectorImpl<Type> &inferredReturnTypes) {
+  return lwe::inferAddOpReturnTypes(ctx, adaptor, inferredReturnTypes);
+}
+
+LogicalResult MulOp::inferReturnTypes(
+    MLIRContext *ctx, std::optional<Location>, MulOp::Adaptor adaptor,
+    SmallVectorImpl<Type> &inferredReturnTypes) {
+  return lwe::inferMulOpReturnTypes(ctx, adaptor, inferredReturnTypes);
+}
+
+LogicalResult MulPlainOp::inferReturnTypes(
+    MLIRContext *ctx, std::optional<Location>, MulPlainOp::Adaptor adaptor,
+    SmallVectorImpl<Type> &inferredReturnTypes) {
+  return lwe::inferMulPlainOpReturnTypes(ctx, adaptor, inferredReturnTypes);
+}
+
+LogicalResult RelinearizeOp::inferReturnTypes(
+    MLIRContext *ctx, std::optional<Location>, RelinearizeOp::Adaptor adaptor,
+    SmallVectorImpl<Type> &inferredReturnTypes) {
+  return lwe::inferRelinearizeOpReturnTypes(ctx, adaptor, inferredReturnTypes);
+}
+
+}  // namespace bgv
+}  // namespace heir
+}  // namespace mlir
