+++ lib/Dialect/LWE/IR/BUILD
@@ -9,7 +9,10 @@ package(
 cc_library(
     name = "Dialect",
     srcs = [
+        "LWEAttributes.cpp",
         "LWEDialect.cpp",
+        "LWEOps.cpp",
+        "LWETypes.cpp",
     ],
     hdrs = [
         "LWEAttributes.h",
@@ -27,6 +30,7 @@ cc_library(
         "@heir//lib/Dialect/ModArith/IR:Dialect",
         "@heir//lib/Dialect/Polynomial/IR:Dialect",
         "@heir//lib/Dialect/RNS/IR:Dialect",
+        "@heir//lib/Utils:APIntUtils",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:IR",
         "@llvm-project//mlir:InferTypeOpInterface",
