+++ lib/Dialect/LWE/IR/LWEAttributes.h
@@ -9,4 +9,24 @@
 #define GET_ATTRDEF_CLASSES
 #include "lib/Dialect/LWE/IR/LWEAttributes.h.inc"
 
+namespace mlir {
+namespace heir {
+namespace lwe {
+
+int64_t getScalingFactorFromEncodingAttr(Attribute encoding);
+
+PlaintextSpaceAttr inferMulOpPlaintextSpaceAttr(MLIRContext *ctx,
+                                                PlaintextSpaceAttr x,
+                                                PlaintextSpaceAttr y);
+
+PlaintextSpaceAttr inferModulusSwitchOrRescaleOpPlaintextSpaceAttr(
+    MLIRContext *ctx, PlaintextSpaceAttr x, APInt dividedModulus);
+
+Attribute getEncodingAttrWithNewScalingFactor(Attribute encoding,
+                                              int64_t newScale);
+
+}  // namespace lwe
+}  // namespace heir
+}  // namespace mlir
+
 #endif  // LIB_DIALECT_LWE_IR_LWEATTRIBUTES_H_
