+++ lib/Dialect/LWE/IR/LWEOps.cpp
@@ -1,0 +1,143 @@
+#include "lib/Dialect/LWE/IR/LWEOps.h"
+
+#include "lib/Dialect/ModArith/IR/ModArithTypes.h"
+#include "llvm/include/llvm/ADT/TypeSwitch.h"    // from @llvm-project
+#include "mlir/include/mlir/IR/TypeUtilities.h"  // from @llvm-project
+
+namespace mlir {
+namespace heir {
+namespace lwe {
+
+//===----------------------------------------------------------------------===//
+// Op verifiers
+//===----------------------------------------------------------------------===//
+
+LogicalResult RMulOp::verify() { return lwe::verifyMulOp(this); }
+
+LogicalResult TrivialEncryptOp::verify() {
+  auto paramsAttr = this->getParamsAttr();
+  auto outParamsAttr = this->getOutput().getType().getLweParams();
+
+  if (paramsAttr != outParamsAttr) {
+    return this->emitOpError()
+           << "lwe_params attr must match on the op and "
+              "the output type, but found op attr "
+           << paramsAttr << " and output type attr " << outParamsAttr;
+  }
+
+  return success();
+}
+
+LogicalResult ReinterpretApplicationDataOp::verify() {
+  auto inputType = getInput().getType();
+  auto outputType = getOutput().getType();
+  if (inputType.getPlaintextSpace() != outputType.getPlaintextSpace() ||
+      inputType.getCiphertextSpace() != outputType.getCiphertextSpace() ||
+      inputType.getKey() != outputType.getKey() ||
+      inputType.getModulusChain() != outputType.getModulusChain()) {
+    return emitOpError()
+           << "the only allowed difference in the input and output are in the "
+              "application_data field, but found input type "
+           << inputType << " and output type " << outputType;
+  }
+
+  return success();
+}
+
+// Verification for RLWE_EncryptOp
+LogicalResult RLWEEncryptOp::verify() {
+  Type keyType = getKey().getType();
+  auto keyRing =
+      llvm::TypeSwitch<Type, mlir::heir::polynomial::RingAttr>(keyType)
+          .Case<lwe::NewLWEPublicKeyType, lwe::NewLWESecretKeyType>(
+              [](auto key) { return key.getRing(); })
+          .Default([](Type) {
+            llvm_unreachable("impossible by type constraints");
+            return nullptr;
+          });
+
+  auto outputRing = getOutput().getType().getCiphertextSpace().getRing();
+  if (outputRing != keyRing) {
+    return emitOpError() << "RLWEEncryptOp input ring do not match. Key ring: "
+                         << keyRing
+                         << ". Output ciphertext ring: " << outputRing << ".";
+  }
+  return success();
+}
+
+// Verify Encoding and Type match
+LogicalResult verifyEncodingAndTypeMatch(mlir::Type type,
+                                         mlir::Attribute encoding) {
+  // En/Decode Ops only allow IntegerOrFloatLike (-> assert not if)
+  assert(getElementTypeOrSelf(type).isIntOrFloat() &&
+         "Encoding Ops only allow IntegerOrFloatLike types");
+
+  // Verification conditions for each encoding we have:
+
+  if (isa<BitFieldEncodingAttr>(encoding)) {
+    // Bit field encodings only work on (scalar/individual) integer
+    return success(type.isInteger());
+  }
+
+  if (isa<UnspecifiedBitFieldEncodingAttr>(encoding)) {
+    // same as BitFieldEncoding
+    return success(type.isInteger());
+  }
+
+  // New LWE Encoding Attr
+
+  if (isa<FullCRTPackingEncodingAttr>(encoding)) {
+    // also supports lists of integers and scalars via replication
+    return success(getElementTypeOrSelf(type).isInteger());
+  }
+
+  if (isa<InverseCanonicalEncodingAttr>(encoding)) {
+    // CKKS-style Encoding should support everything
+    // (ints via cast to float/double, scalars via replication)
+    return success();
+  }
+
+  // This code should never be hit unless we added an encoding and forgot to
+  // update this function. Assert(false) for DEBUG, return failure for NDEBUG.
+  encoding.dump();
+  assert(false && "Encoding not handled in encode/decode verifier.");
+  return failure();
+}
+
+LogicalResult EncodeOp::verify() {
+  return verifyEncodingAndTypeMatch(getInput().getType(), getEncoding());
+}
+
+LogicalResult RLWEEncodeOp::verify() {
+  return verifyEncodingAndTypeMatch(getInput().getType(), getEncoding());
+}
+
+LogicalResult RLWEDecodeOp::verify() {
+  return verifyEncodingAndTypeMatch(getResult().getType(), getEncoding());
+}
+
+//===----------------------------------------------------------------------===//
+// Op type inference.
+//===----------------------------------------------------------------------===//
+
+LogicalResult RAddOp::inferReturnTypes(
+    MLIRContext* ctx, std::optional<Location>, RAddOp::Adaptor adaptor,
+    SmallVectorImpl<Type>& inferredReturnTypes) {
+  return lwe::inferAddOpReturnTypes(ctx, adaptor, inferredReturnTypes);
+}
+
+LogicalResult RSubOp::inferReturnTypes(
+    MLIRContext* ctx, std::optional<Location>, RSubOp::Adaptor adaptor,
+    SmallVectorImpl<Type>& inferredReturnTypes) {
+  return lwe::inferAddOpReturnTypes(ctx, adaptor, inferredReturnTypes);
+}
+
+LogicalResult RMulOp::inferReturnTypes(
+    MLIRContext* ctx, std::optional<Location>, RMulOp::Adaptor adaptor,
+    SmallVectorImpl<Type>& inferredReturnTypes) {
+  return lwe::inferMulOpReturnTypes(ctx, adaptor, inferredReturnTypes);
+}
+
+}  // namespace lwe
+}  // namespace heir
+}  // namespace mlir
