+++ lib/Dialect/LWE/IR/LWETraits.td
@@ -7,4 +7,8 @@ def SameOperandsAndResultRings:  NativeOpTrait<"SameOperandsAndResultRings"> {
   let cppNamespace = "::mlir::heir::lwe";
 }
 
+def SameOperandsAndResultPlaintextTypes:  NativeOpTrait<"SameOperandsAndResultPlaintextTypes"> {
+  let cppNamespace = "::mlir::heir::lwe";
+}
+
 #endif // LIB_DIALECT_LWE_IR_LWETRAITS_TD_
