+++ lib/Dialect/Openfhe/IR/BUILD
@@ -10,6 +10,7 @@ cc_library(
     name = "Dialect",
     srcs = [
         "OpenfheDialect.cpp",
+        "OpenfheOps.cpp",
     ],
     hdrs = [
         "OpenfheDialect.h",
@@ -20,7 +21,6 @@ cc_library(
         ":dialect_inc_gen",
         ":ops_inc_gen",
         ":types_inc_gen",
-        "@heir//lib/Dialect:FHEHelpers",
         "@heir//lib/Dialect/LWE/IR:Dialect",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:IR",
