+++ lib/Dialect/Openfhe/IR/OpenfheOps.td
@@ -209,7 +209,7 @@ def SubPlainOp : Openfhe_Op<"sub_plain",[
   let results = (outs NewLWECiphertext:$output);
 }
 
-def MulOp : Openfhe_BinaryOp<"mul", [AllTypesMatch<["lhs", "rhs", "output"]>]> { let summary = "OpenFHE mul operation of two ciphertexts with relinearization."; }
+def MulOp : Openfhe_BinaryOp<"mul"> { let summary = "OpenFHE mul operation of two ciphertexts with relinearization."; }
 
 def MulNoRelinOp : Openfhe_Op<"mul_no_relin", [Pure, SameOperandsAndResultRings]> {
   let summary = "OpenFHE mul operation of two ciphertexts without relinearization.";
@@ -222,8 +222,7 @@ def MulNoRelinOp : Openfhe_Op<"mul_no_relin", [Pure, SameOperandsAndResultRings]
 }
 
 def MulPlainOp : Openfhe_Op<"mul_plain",[
-    Pure,
-    AllTypesMatch<["ciphertext", "output"]>
+    Pure
 ]> {
   let summary = "OpenFHE mul operation of a ciphertext and a plaintext.";
   let arguments = (ins
@@ -252,7 +251,14 @@ def SquareOp : Openfhe_UnaryOp<"square"> { let summary = "OpenFHE square operati
 def RelinOp : Openfhe_UnaryTypeSwitchOp<"relin"> { let summary = "OpenFHE relinearize operation of a ciphertext."; }
 
 def ModReduceOp : Openfhe_UnaryTypeSwitchOp<"mod_reduce"> { let summary = "OpenFHE mod_reduce operation of a ciphertext. (used only for BGV/CKKS)"; }
-def LevelReduceOp : Openfhe_UnaryTypeSwitchOp<"level_reduce"> { let summary = "OpenFHE level_reduce operation of a ciphertext."; }
+def LevelReduceOp : Openfhe_UnaryTypeSwitchOp<"level_reduce"> {
+  let summary = "OpenFHE level_reduce operation of a ciphertext.";
+  let arguments = (ins
+    Openfhe_CryptoContext:$cryptoContext,
+    NewLWECiphertext:$ciphertext,
+    DefaultValuedAttr<I64Attr, "1">:$levelToDrop
+  );
+}
 
 def RotOp : Openfhe_Op<"rot",[
   Pure,
